--- server/src/main/java/nextmainfocus/security/entity/Account.java
@@ -1,45 +1,0 @@
-package nextmainfocus.security.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-public class Account {
-    @Getter
-    @Setter
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    @Getter
-    @Setter
-    private String username;
-    @Getter
-    @Setter
-    private String password;
-    @Getter
-    @Setter
-    private String role;
-
-// methods below are from Spring Security and not used for now
-
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    public boolean isEnabled() {
-        return true;
-    }
-}
