--- server/src/main/java/nextmainfocus/tasks_list/entity/Task.java
@@ -1,37 +1,0 @@
-package nextmainfocus.tasks_list.entity;
-
-import java.time.LocalDateTime;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
-import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-public class Task {
-    @Getter
-    @Setter
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    @Getter
-    @Setter
-    private String description;
-    @Getter
-    @Setter
-    private boolean isDone;
-
-    @Getter
-    @Setter
-    @JsonDeserialize(using = LocalDateTimeDeserializer.class)
-    @JsonSerialize(using = LocalDateTimeSerializer.class)
-    private LocalDateTime dateTime;
-}
