--- server/src/main/resources/old-application.properties
@@ -1,14 +1,0 @@
-spring.datasource.url=jdbc:mariadb://localhost:3306/todolist
-spring.datasource.username=root
-spring.datasource.password=example
-spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
-
-spring.jpa.show-sql=true
-# use 'update' if this is the first time to run the app and database is empty
-# https://stackoverflow.com/questions/56283839/hibernate-caused-by-java-sql-sqlexception-table-view-users-already-exists-i
-#spring.jpa.hibernate.ddl-auto=update
-spring.jpa.hibernate.ddl-auto=none
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
-
-# Client URL
-client.url=*
