+++ settings.gradle
@@ -4,4 +4,4 @@ pluginManagement {
 		gradlePluginPortal()
 	}
 }
-rootProject.name = 'mainService'
+rootProject.name = 'nextmainfocus'
