+++ src/main/java/nextmainfocus/tasks_list/Sorting.java
@@ -1,5 +1,5 @@
 package nextmainfocus.tasks_list;
 
 public enum Sorting {
-    ASC, DESC
+	ASC, DESC
 }
